Resolved conflict

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -28,7 +28,7 @@ public class SalesManager {
     public int averageValue() {
         int sum = 0;
         int i = -2;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale != min() && sale != max())
                 sum += sale;
             i++;
